+++ src/main/java/frc/robot/Constants.java
@@ -7,8 +7,8 @@ import edu.wpi.first.math.geometry.Transform3d;
 import edu.wpi.first.math.geometry.Translation3d;
 import edu.wpi.first.math.util.Units;
 import edu.wpi.first.wpilibj.XboxController;
+import frc.robot.pioneersLib.VisionUtil.CameraResolution;
 import frc.robot.pioneersLib.bumSwerve.SwerveModule;
-import frc.robot.pioneersLib.bumSwerve.VisionUtil.CameraResolution;
 import frc.robot.pioneersLib.bumSwerve.Gyro.SwerveGyroIO;
 import frc.robot.pioneersLib.bumSwerve.Gyro.SwerveGyroIONavX;
 import frc.robot.pioneersLib.bumSwerve.Gyro.SwerveGyroIOSim;
